Доработан контроллер избранного

+++ app/Http/Controllers/FavoriteController.php
@@ -7,6 +7,7 @@ use App\Http\Responses\FailResponse;
 use App\Http\Responses\SuccessResponse;
 use App\Models\Film;
 use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpFoundation\Response;
 
 class FavoriteController extends Controller
 {
@@ -18,9 +19,9 @@ class FavoriteController extends Controller
     public function index(): BaseResponse
     {
         try {
-            /** @var \App\Models\User|null $user */
+            /** @var User $user */
             $user = Auth::user();
-            $favoriteFilms = $user->favoriteFilms()->get();
+            $favoriteFilms = $user->favoriteFilms()->orderBy('created_at', 'desc')->get();
 
             return new SuccessResponse($favoriteFilms);
         } catch (\Exception $e) {
@@ -36,8 +37,13 @@ class FavoriteController extends Controller
     public function store(Film $film): BaseResponse
     {
         try {
-            /** @var \App\Models\User|null $user */
+            /** @var User $user */
             $user = Auth::user();
+
+            if ($user->favoriteFilms()->where('film_id', $film->id)->exists()) {
+                return new FailResponse('Фильм уже в избранном', Response::HTTP_UNPROCESSABLE_ENTITY);
+            }
+
             $user->favoriteFilms()->attach($film);
 
             return new SuccessResponse();
@@ -54,8 +60,13 @@ class FavoriteController extends Controller
     public function destroy(Film $film): BaseResponse
     {
         try {
-            /** @var \App\Models\User|null $user */
+            /** @var User $user */
             $user = Auth::user();
+
+            if (!$user->favoriteFilms()->where('film_id', $film->id)->exists()) {
+                return new FailResponse('Фильм не найден в избранном', Response::HTTP_UNPROCESSABLE_ENTITY);
+            }
+
             $user->favoriteFilms()->detach($film);
 
             return new SuccessResponse();
